refactor: simplify scancode conversion using new Resource functions
